[ASCT-133] Use vcpkg to manage dependencies (#17)
# Use vcpkg to manage dependencies

## Description
In order to simplify the build process and make it easy for developers
to check out and build ASTL without hunting down dependencies
themselves, we'll use a package manager to download and build them for
us. Making this explicit in the vcpkg.json manifest could enable
dependency updating automation at some point if/when dependabot supports
C++.

I evaluated vcpkg against some other options that i've dealt with in
past projects:

0. do nothing, require developers to download, build, and install
libraries and set up paths themselves. Too onerous for developers
1. copy dependencies into repo, add build scripts. Too much wasteful
cloning, dependencies get out of date, maintenance effort on build
scripts
2. cmake FetchContent - actually pretty great for simple header-only
libraries. Doesn't seem as fully featured as conan or vcpkg
3. conan - seems similar to vcpkg in capabilities. My experience with it
in the past was slightly frustrating, and it doesn't seem to be quite as
popular or well supported. but this could be a skill issue. If we ever
need some complex dependencies, more sophisticated dependency caching,
or want to set up our own package repo for dependencies, conan could be
a viable alternative

How this works:
vcpkg extends cmake's find_package in a way that's basically magic from
our perspective. The only additional requirement is the user needs to
install vcpkg up front. I added an error message at the top of the
cmakelists with instructions on how to do this.

vcpkg should come installed by default on github cloud runners, but i
added a workflow step to check for it too, since we might run this on
self-hosted at some point.

I threw in a CMakePresets.json too. This is was originally to define the
cmake tool to use vcpkg, but that turned out to be a hard-to-read error
if unavailable. I left this in place as it's a good place to put config
data if we want to in the future.

The pilot dependency is Catch2 - a unit testing framework.

All in all, I'm pretty happy with the experience. I've never used vcpkg,
and my cmake skills are not the best, but I got a unit test application
off the ground in about 2 hours with this.

Fixes #
* [ASCT-133](https://jira.arm.com/browse/ASCT-133)
* #14
* #8

## Customer Impact

Developers will need to install vcpkg before building ASTL.

## Performance Impact

Build time is increased, because we're adding a step to acquire Catch2 -
a 3rd party dependency.

## Type of change

Please delete options that are not relevant.
- [X] This change requires a documentation update
- [X] This changes requires a version update
The changes i think are needed because the instructions to build have
changed

## How Has This Been Tested?

- [X] Build on astra1 ubuntu with no vcpkg - confirmed relevant
instructions are given to set it up
- [X] Build on astra1 successfully, including Catch2 tests
- [X] First unit test failing by modifying implementation of
AstlCollector::test()
- [X] First unit test passing by using original ::Test behavior.

**Test Configuration**:
* OS name: Ubuntu
* OS version:
* Kernel version:
* Hardware name:
* Hardware stepping:
* Firmware version:
* Machine name: astra1

## Checklist:

- [x] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [X] I have checked my code and corrected any misspellings

Ossmosis-Source-Commit: a32ab0675f0086693d59fe5710ed5fb2652627ed