Add placeholder tests for C wrapper layer (#23)
# Implement placeholder tests to increase baseline code coverage

## Description

Assert that all unimplemented functions return the unimplemented error
code.
Also resolve issue #22 
The main purpose here is to get our code coverage up to a high threshold
so we don't regress on it in the future, with the help of CodeClimate.

Fixes #22

## Customer Impact

None
## Performance Impact

None
## Type of change

Please delete options that are not relevant.

- [X] Bug fix (non-breaking change which fixes an issue)

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

- [X] On astra1 platform, generating coverage and viewing html output
with gcovr
- [X] In Github actions, watching CodeClimate's reports added as checks.

**Test Configuration**:
* OS name: ubuntu
* OS version:
* Kernel version:
* Hardware name:
* Hardware stepping:
* Firmware version:
* Machine name: astra1

## Checklist:

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [X] I have checked my code and corrected any misspellings