[ASTL-50] Header utility for logging and writing output files (#41)
## Description

Logging utility for writing formatted log messages with severity level
and source location to the console or to a file. The utility can also be
used to log preformated text strings for output generation

Fixes # (ASTL-50)

## Customer Impact

Utility would be used as part of the development process and will also
be used to log errors and critical failures to the user

## Performance Impact

None known

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Interface change (fix or feature that would cause the user facing
interface to change)
- [ ] This change requires a documentation update
- [ ] This changes requires a version update

## How Has This Been Tested?

Manual tests with combination of console and file logging as well as
environement variables
TODO: unit testing in separate patch.

**Test Configuration**:
* OS name: Ubuntu
* OS version: 24.04
* Kernel version: 6.12.10
* Machine name: asct-astra2

## Checklist:

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [X] I have checked my code and corrected any misspellings

---------

Co-authored-by: Faycal Benmlih <fayssal.benmlih@arm.com>