[NO-JIRA] Fix up some warnings due to bad test definitions (#105)
## Description

The trompeloeil REQUIRE_CALL macro has an optional WITH() macro. It's
contents should be an expression that evaluates to a boolean. I was
giving it two conditions with a comma, which means the first result is
ignored and the second expression is treated as the condition.

WRONG:
```
.WITH(_1 == "file", _2 == "1")
```

CORRECT:
```
.WITH(_1 == "file" && _2 == "1")
// or
.WITH(_1=="file")
.WITH(_2=="1")
```

Fixes # (issue)
Several warnings about unused boolean result.

## Checklist:

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [X] I have checked my code and corrected any misspellings