[ASTL-108] Add an AstlValue variant utility class (#108)
## Description

This is a C++ typesafe union of the astl_value_t type in the C API. The
intent is to make computing averages, min, max, and formatting to
strings easier for astl_value_t types by isolating all that work in the
astl_value.hpp file.

Fixes # [ASTL-108]

## Customer Impact

N/A

## Performance Impact

I ran a [quickbench](https://quick-bench.com/) comparison and found that
a 100% cpu bound routine computing average values of astl_value_t and
astl::AstlValue resulting in a 1.4x slowdown with astl::AstlValue.

![image](https://github.com/user-attachments/assets/3b154728-1699-46a7-b06a-7a82ff4cb425)
It's a little slower, but it's also less code: about 6 lines for the
variant case, 140 lines for the tagged union case. (there are a lot of
cases to cover!). I also doubt that summing values is going to be our
program bottleneck, so I'm not too concerned about the performance delta
here.


[quickbench_notes.txt](https://github.com/user-attachments/files/20760828/quickbench_notes.txt)

## Type of change

- [X} Internal infrastructure improvement

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

- [X] Added CI tests showing that addition, division, and printing work
for a few type cases.
- [X] Performance benchmark testing with quickbench to show we're not
too out of whack


## Checklist:

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [X] I have checked my code and corrected any misspellings