[ASTL-107] Added VCPKG to lint flow to manage dependencies (#118)
## Description

The lint flow has a bug where the dependencies are not cloned. These
dependencies are currently managed by vcpkg. This PR adds VCPKG
dependency management to the lint flow.

Since VCPKG is run as part of the configure step, there is no need for
an explicit fast .in render flow. We can still keep the separate
render_templates cmake module though.

## Customer Impact

Bug fix with CICD

## Performance Impact

Lint flow is slower due to VCPKG, but it's not on the critical path so
overall CICD latency will be unaffected.

## Type of change

- [X] Bug fix (non-breaking change which fixes an issue)
- [X] Testing, Verification, or CICD change

## How Has This Been Tested?

CICD Must Pass

**Test Configuration**:
* OS name: Ubuntu and MacOS