[NO-JIRA] Enabled class diagrams and function call graphs with Doxygen (#149)
## Description

Brought up doxygen to generate pretty diagrams automatically from ASTL
source code.

## Customer Impact

This will probably be mostly for internal documentation, but we could
potentially use this to generate external documentation as well in the
future.

## Performance Impact

None

## Type of change

- [X] New feature (non-breaking change which adds functionality)
- [X] Testing, Verification, or CICD change
- [X] Documentation

## How Has This Been Tested?

Tested locally in my terminal. This is not part of CICD yet (and may not
need to be)

**Test Configuration**:
* OS name: Ubuntu
* OS version: 22.04

---------

Co-authored-by: qltysh[bot] <168846912+qltysh[bot]@users.noreply.github.com>

Ossmosis-Source-Commit: 273b016d88bdb44de94342cdceae5ad1c889a30d