[NO-JIRA] Doxygen now outputs to pre-existing doc folder (#152)
## Description

I wasn't aware of the doc folder when I initially brought up Doxygen.
It's rendered outputs are now saved there

## Customer Impact

For now, the Doxygen outputs are mostly there for internal development,
but they could be used for customer facing docs in the future if we
wanted.

## Performance Impact

None

## Type of change

- [X] Testing, Verification, or CICD change

## How Has This Been Tested?

Tested locally.  This feature does not need to be part of CICD.

## Test Configuration

- OS name: Ubuntu
- OS version: 22.04

Ossmosis-Source-Commit: 6ade91cae1e169b34df2dd6174fce7c052efa81c