Fix stringify for REQUIRE asserts for astl_status_code (#163)
## Description

Failed tests will now report the enum values as labels:
```
Failure:
    REQUIRE_THAT(astlDummyFunction() == ASTL_STATUS_SUCCESS)
with expansion:
    ASTL_STATUS_INTERNAL_ERROR == ASTL_STATUS_BAD_ARGUMENT
```

as opposed to the raw numeric values:
```
with expansion:
    42 == 32
```

Also using magic_enum to simplify the way the implementation of
astlStatusString. This required that we fit the status codes to
stringify into the range of -127 to 127. There are supposedly ways to
configure magic_enum regarding this range, or customize the range for a
particular enum type, but I couldn't get any of them working, so i
changed ASTL_STATUS_INTERNAL_ERROR from 200 to 100.

Also added a justfile as a convenient command runner, a tip I'm
borrowing from Zach

Fixes #96 #16

## Customer Impact

ASTL_STATUS_INTERNAL_ERROR changes value from 200 to 100

## Performance Impact

None

## Type of change

- [X] Bug fix (non-breaking change which fixes an issue)
- [X] Testing, Verification, or CICD change
- [X] Interface change (fix or feature that would cause the user facing
interface to change)

## How Has This Been Tested?

Run all tests on astra2

## Test Configuration

- OS name:
- OS version:
- Kernel version:
- Hardware name: NA
- Hardware stepping: NA
- Firmware version: NA
- Machine name: NA

## Checklist

- [X] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [X] I have checked my code and corrected any misspellings

Ossmosis-Source-Commit: 1b57176c814f660194fd233006d16262e1e4464f