[ASTL-137] configuration json specifies metrics (#165)
## Description

Replace hard-coded static metric configuration with json-configurable
metrics.
For instance, the metric:
```
    "SoC Power": {
      "description": "SoC Power Consumption in Watts",
      "register": "ENERGY_COUNTER",
      "unit": "W",
      "metric_type": "rate",
      "collector": "scmi"
    }
```
added to the top level library configuration file will tell metric
manager to collect a RateMetric through SCMI on registers with the name
'ENERGY_COUNTER'. In the sample scmi platform specification, there is
one 'ENERGY_COUNTER' per Application Processor (AP) target.

One change to how MetricConfigs work is that rather than taking a
collection of data events to read for a single metric, a MetricConfig
now owns a `ScmiTargetToDataEventIdMap` - basically a map from target to
data event ID on that target.

In this example, since ENERGY_COUNTER is defined in the sample scmi spec
as
```
    "ENERGY_COUNTER": {
          "instance_id": 0,
          "local_id": 55,
          "offset": "0x00F0",
          "name": "AP0_ENERGY_COUNTER",
          "line_addr": "0x740080F0",
          "value_addr": "0x740080F8",
          "de_id": "0x00000037"
``` 
on AP0, and similarly, but with de_id 0x00010037 for AP1, we'd have one
'SoC Power' MetricConfig with a map of AP0 -> 0x37, AP1 -> 0x10037.

Fixes # ASTL-137

## Customer Impact
Faster iteration on demos!
We ought to be able to create and demonstrate scenarios just by
extending our 2 json config files and creating a .csv file for canned
data in MockSysFs, with no need to recompile.

## Performance Impact

Please describe any relevant performance impact of this change. This can
be positive or negative impact. How did you characterize/test the
performance impact?

## Type of change

- [X] New feature (non-breaking change which adds functionality)
- [X] Interface change - configuration file schema extended
- [X] This change requires a documentation update

## How Has This Been Tested?

Running the demo.sh locally on astra2, and in CI

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules
- [x] I have checked my code and corrected any misspellings