[ASTL-152] Add `BuildCollectors` (#175)
## Description

The `collector_builder.cpp` translation unit should know how to
instantiate all various collector types (we expect the variety of
collector types to be an extension point). Ideally, other components of
the library can deal only with the ICollector interface, not specifics
like ScmiSysfsCollector.

[ASTL-148] This also removes the ForceTargetToCollectorMap method from
CollectorManager. The set of collectors in a collector manager is a ctor
parameter and isn't expected to change on the fly once collectorManager
is instantiated.

Fixes # ASTL-152  ASTL-148

## Customer Impact

None

## Performance Impact

None

## Type of change

- [X] Internal Interface change

## How Has This Been Tested?
CI, astra2 unit tests and demo.

## Checklist

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [X] I have checked my code and corrected any misspellings

Ossmosis-Source-Commit: 62d7f73cd6ef0d6f8ded24c905839d6d3da95e2c