[ASTL-153] Add output checks on demo.sh to CI (#187)
## Description

Confirm that the last output message, from astlGetMetricSamplesOnTarget
successfully prints to the console.

sample_test.cpp can still return a 0 error code for failures, since it
runs as part of the integration unit tests without mocksysfs.

But when demo.sh adds mocksysfs to the mix, we expect more valid output
strings on the console.

Fixes # ASTL-153

## Customer Impact

sample code is more likely to be consistently functional, and demo.sh is
more certain to work on any commit on the main branch.

## Performance Impact

None

## Type of change

- [X] Testing, Verification, or CICD change

## How Has This Been Tested?

First, by checking in a breaking change to sample_test, to ensure that
CI will fail.

https://github.com/Arm-Debug/ASTL/actions/runs/17106681918/job/48517332924#step:18:3
Second, by reverting the bug in sample_test to see a positive outcome.

https://github.com/Arm-Debug/ASTL/actions/runs/17107216189/job/48519147043#step:18:376

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules
- [x] I have checked my code and corrected any misspellings

Ossmosis-Source-Commit: 674227f6080812567d436df8ff7347287ae5c83d