[ASTL-140] Add SensorsCollector implementation (#193)
## Description [ASTL-140] Implement a new Collector type on libsensors

Built on the libsensors library in Linux.
https://linux.die.net/man/3/libsensors
This library provides basic metrics like temperature sensors, power
consumption, and fan speed.
It might prove useful for ASTL POC in systems without SCMI.

Unit tests are built with a link seam, replacing the libsensors
implementation with a shim to a mock instance.

Note that this _only_ adds the collector class definition, it doesn't
integrate an instance of the collector, or add metrics that are capable
of interfacing with libsensors.

These tickets should handle the rest:
[ASTL-138] - ASTL: HWMON - extend topology manager to identify HWMON
sysfs target(s)
[ASTL-139] - ASTL: HWMON - extend configuration manager to enumerate
HWMON sysfs metrics

Fixes # [ASTL-140]

## Customer Impact
Additional metrics for power and temperature available outside of SCMI
(once config, topology components for Libsensors are in place)

## Performance Impact

None

## Type of change

- [X] New feature (non-breaking change which adds functionality)

## How Has This Been Tested?

Built on astra2 and my macbook (to ensure things compile without the
sensors.h library present).
Minimal unit tests added with a mock implementation of the library.

## Test Configuration

- Machine name: astra2

## Checklist

- [X] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [x] I have checked my code and corrected any misspellings