[ASTL-144] Implement SCMI Topology Plugin (#190)
## Description

Implements an actual functional SCMI plugin for the topology manager
which tries to detect if a working SCMI interface is available on the
platform.

## Customer Impact

ASTL can automatically detect the existance of a SCMI interface.

## Performance Impact

Adds some small overhead and filesystem operations.

## Type of change

- [X] New feature (non-breaking change which adds functionality).
Well... this PR kinda broke on of the unit tests, but this was only
because the test was expecting fake data from some placeholder code.

## How Has This Been Tested?

* Unit tests must pass.
* Topology manager must initialize properly when using MockSyfs.
* Note: Needed to manually test against an unmerged version of MockSysFS
#180 which adds de_implementation_version.

## Test Configuration

- OS name: Ubuntu
- OS version: 24.04

---------

Co-authored-by: Brendan Long <brendan.long@arm.com>
Co-authored-by: Brendan Long <bccbrendan@gmail.com>