[ASTL-162] Implement output manager, buffer output and processed samples sink
## Description

Create metric manager and buffer output writer along with a processed
sample sink from metric instances to the orchestrator. The sinked
samples are saved in a target map to maps of metrics to sample spans.
The output manager uses the target and metric to key the appropriate
span of samples to output.

## Customer Impact

As part of the patch, the interface was simplified by removing the
GetAll samples functions and sample structures no longer have a target
handle and metric handle fields. The user is expected to get processed
samples for specific metrics for specific targets and will have to call
GetSamples function multiple times and do book keeping to get all of the
samples for all metrics on all of the targets

## Performance Impact

writing processed samples to buffer is done at the end when the
collection is already do so no expected performance impact while user
workload is running.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Testing, Verification, or CICD change
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [X] Interface change (fix or feature that would cause the user facing
interface to change)
- [ ] This change requires a documentation update
- [ ] This changes requires a version update

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

- [X] Mocking classes for output manager and buffer output added
- [X] Unite tests for output manager and buffer output added

## Test Configuration
- Machine name: ASTRA2

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules
- [x] I have checked my code and corrected any misspellings

Ossmosis-Source-Commit: a77d5f02efb02f979e15a333d8c8a7704a0fc04a