[ASTL-171] refactor protocol telemetry functions (#213)
## Description

Refactor BuildTelemetryTree into 3 separate functions for top level
files, des directory, and groups directory.

Refactor GetTelemetryFile into a map instead of if conditionals.

Previously each data event stored a raw pointer to its group. It was
then changed to only store a group id.
Because of this change, we don't need to build a list of data events -
just check that if we are handling a group file, the group ids match.

Fixes # (issue)

## How Has This Been Tested?

ctests

## Test Configuration

- OS name: Ubuntu
- OS version: 22.04