[ASTL-175] Residency Metric Processed sample sinking. (#218)
## Description
The residency Metric type is a specialized Metric type tracking the
residency of various states. While invoking the GetSamples API the
Orchestrator/Output Manager need to know the order of the samples it
receive. The metric instance has to make sure the samples are sinked in
a deterministic order as reported.

Fixes # (issue)

## Performance Impact

This feature reduces the memory footprint of Residency metric samples.
The state information of each samples are encoded in the Sample sinking
order.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Testing, Verification, or CICD change
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Interface change (fix or feature that would cause the user facing
interface to change)
- [ ] This change requires a documentation update
- [ ] This changes requires a version update

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

- [x] Unit testing
- [x] CI/CD workflow

## Test Configuration

- OS name:
- OS version:
- Kernel version:
- Hardware name: NA
- Hardware stepping: NA
- Firmware version: NA
- Machine name: NA

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules
- [x] I have checked my code and corrected any misspellings

Ossmosis-Source-Commit: c62d024e2585fd8c2d4705f5756d9c0df1ce6afb