[ASCT-296] Switch to shared organization-level blackduck API token (#221)
## Description

Switched from a random team member's blackduck API key to a proper
org-level API key which is associated with a bot account
(svc_its_ci@arm.com).

## Customer Impact

Mostly internal, but this improves stability of our verification flows
and reduces the risk of failure which could delay a release.

## Performance Impact

None

## Type of change

- [X] Bug fix (non-breaking change which fixes an issue)
- [X] Testing, Verification, or CICD change

## How Has This Been Tested?

CI/CD must pass. In particular, the Blackduck CI must run with the new
API key. Since blackduck is not part of the PR-checkin flow, I will need
to run this manually and verify by inspection.