[ASTL-227] Refactor Orchestrator (#238)
## Description

**Motivation:**
Our configuration json files are meant to be implementation details for
us to modify, and shouldn't be modified by the customer. The customer
doesn't even need to understand the contents of these files.

**Changes:**

1. Separate astlInitialize() into two functions.
a. One C-style function callable by the customer. Eventually, it will
not have a parameter to specify a config file path, but this change will
be finished in a future PR (ASTL-183).
b. The other is a C++-style function similar to the other builder
functions which already exist. The bulk of the old astlInitialize() will
live here. The c-style function will basically be a wrapper for this
function.
2. As mentioned above, refactored astlInitialize() into a Builder
function to be more consistent with the various managers. This is mostly
a rename, the structure was already set up this way.
3. Moved a few files into a new "orchestrator" directory and renamed
files with generic names to orchestrator names. Again, this is mostly to
improve consistency with the rest of the codebase.
4. This two-function structure allows us to perform dependency injection
in our unit tests. The C++-style function gives us all the control that
astlInitialze() did and unit tests can call this. The C-style function
is just a thin wrapper, so will only need minimal testing.

Still TODO (ASTL-183) - Actually remove the configuration file path from
the C-style interface. This will be done in a separate, future PR. Will
probably need to write an extra helper function to get the path of the
astl .so file rather than the client executable.

## Customer Impact

Customer will not need to be aware of our internal configuration file to
use ASTL.

## Performance Impact

None

## Type of change

- [X] Interface change (fix or feature that would cause the user facing
interface to change)
- [X] Refactor

## How Has This Been Tested?

CICD Must Pass.

## Test Configuration

- OS name: Ubuntu
- OS version: 22.04 24.04