[ASTL-226] Resolve formatting discrepancy in .md files (#248)
## Description

`scripts/format.sh` seemed to be having inconsistent results on some
markdown files. Sometimes adding blank lines after enumerated items,
sometimes deleting them. This just added a bit of noise and friction in
pull requests.

Turns out to be a difference in the default version of `prettier` - the
markdown formatter, used on MacOs (3.5.3) versus astra2's ubuntu
(3.6.2).

we can just pin the version to 3.6.2 in the .qlty/qlty.toml to resolve
this. I'm also deleting or commenting on the use of some qlty plugins,
so we know what we're scanning for (and no longer scanning for
misconfigured K8s clusters).

Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

Fixes # ASTL-226

## Customer Impact

None

## Performance Impact

slightly faster CI as we skip some unnecessary checks (`checkov` plugin
for qlty)
## Type of change

- [X] Testing, Verification, or CICD change

## How Has This Been Tested?

Run `./scripts/format.sh` on astra2 and on MacOs.


## Checklist

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules
- [X] I have checked my code and corrected any misspellings