[ASTL-183] Remove astlInitialize references from README.md and Cython (#259)
## Description

Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

Fixes # (issue)

## Customer Impact

Please describe any customer facing impact of this change. This can be
positive or negative impact.

## Performance Impact

Removed old references to astlInitialize() from documentation after the
function was refactored out. The example in README.md was completely
irrelevent and I didn't feel a new example was needed. Grepped through
the rest of the codebase and fixed other references.

## Type of change

- [X] Documentation Update

## How Has This Been Tested?

CICD must pass, in particular the formatting checks.

## Test Configuration

- OS name: Ubuntu
- OS version: 22.04, 24.05

## Checklist
 - [x] My code follows the style guidelines of this project
 - [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
 - [x] I have made corresponding changes to the documentation
 - [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
 - [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules
 - [x] I have checked my code and corrected any misspellings